+++ src/com/intellij/idea/plugin/hybris/acl/file/actions/AclFileCreateAction.kt
@@ -19,8 +19,8 @@ package com.intellij.idea.plugin.hybris.acl.file.actions
 
 import com.intellij.ide.actions.CreateFileFromTemplateAction
 import com.intellij.ide.actions.CreateFileFromTemplateDialog
-import com.intellij.idea.plugin.hybris.actions.ActionUtils
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
+import com.intellij.idea.plugin.hybris.util.isHybrisProject
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
@@ -42,7 +42,7 @@ class AclFileCreateAction : CreateFileFromTemplateAction(NEW_FILE, "", HybrisIco
         super.postProcess(createdElement, templateName, customProperties)
     }
 
-    override fun isAvailable(dataContext: DataContext) = super.isAvailable(dataContext) && ActionUtils.isHybrisContext(dataContext)
+    override fun isAvailable(dataContext: DataContext) = super.isAvailable(dataContext) && dataContext.isHybrisProject
     override fun getActionName(directory: PsiDirectory, newName: String, templateName: String) = NEW_FILE
     override fun getDefaultTemplateProperty() = null
     override fun hashCode() = javaClass.hashCode()