+++ src/com/intellij/idea/plugin/hybris/diagram/typeSystem/TSDiagramElementManager.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,19 +20,17 @@ package com.intellij.idea.plugin.hybris.diagram.typeSystem
 
 import com.intellij.diagram.AbstractDiagramElementManager
 import com.intellij.diagram.DiagramBuilder
-import com.intellij.idea.plugin.hybris.actions.ActionUtils
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.diagram.typeSystem.node.graph.*
 import com.intellij.idea.plugin.hybris.system.type.meta.model.TSMetaRelation
+import com.intellij.idea.plugin.hybris.util.ifHybrisProject
 import com.intellij.openapi.actionSystem.DataContext
 import com.intellij.ui.SimpleColoredText
 import com.intellij.ui.SimpleTextAttributes
 
 class TSDiagramElementManager : AbstractDiagramElementManager<TSGraphNode>() {
 
-    override fun findInDataContext(dataContext: DataContext) = if (ActionUtils.isHybrisContext(dataContext))
-        TSGraphNodeRoot()
-    else null
+    override fun findInDataContext(dataContext: DataContext) = dataContext.ifHybrisProject { TSGraphNodeRoot() }
 
     override fun isAcceptableAsNode(item: Any?) = item is TSGraphNode
     override fun getElementTitle(node: TSGraphNode?) = node?.name